Feature "Take a photo for the post" is Done
Re-shaped the dialog view
Changed the layout to Scrollview Layout, this way the
user will be able to scroll up and down when the image
is long.

I think it is okay to keep the two buttons "Camera" and
"Upload" after the photo is in the imageView. The user
can re-take or re-upload if they are not satisfied with
the photo.